refactor interface files
