+++ src/types/rating.ts
@@ -1,0 +1,4 @@
+export interface Rating {
+  Source: string;
+  Value: string;
+}
