+++ src/types/searchResult.ts
@@ -1,52 +1,4 @@
-export interface Item {
-  Title: string;
-  Year: string;
-  Type: string;
-  Plot?: string;
-  imdbID: string;
-}
-
-export interface FullItem extends Item {
-  Rated: string;
-  Released: string;
-  Runtime: string;
-  Genre: string;
-  Director: string;
-  Writer: string;
-  Actors: string;
-  Plot: string;
-  Language: string;
-  Country: string;
-  Awards: string;
-  Poster: string;
-  Ratings: Rating[];
-  Metascore: string;
-  imdbRating: string;
-  imdbVotes: string;
-  DVD: string;
-  BoxOffice: string;
-  Production: string;
-  Website: string;
-  Response: string;
-}
-
-export interface Rating {
-  Source: string;
-  Value: string;
-}
-
-export interface FormattedItem {
-  Title: string;
-  Year: string;
-  Type: string;
-  Plot?: string;
-  'IMDb ID': string;
-}
-
-export interface FormattedAverageSeason {
-  [key: string]: string;
-  'IMDb score': string;
-}
+import { FormattedItem } from './item';
 
 export enum SearchResultType {
   Movies = 'movie',
