+++ src/types/season.ts
@@ -1,0 +1,4 @@
+export interface FormattedAverageSeason {
+  [key: string]: string;
+  'IMDb score': string;
+}
